Solve part two of PacketDecoder

+++ src/main/java/day/_16/LiteralPacket.java
@@ -2,9 +2,9 @@ package day._16;
 
 public class LiteralPacket extends Packet {
 
-    protected final int value;
+    protected final long value;
 
-    public LiteralPacket(int version, int typeId, int value) {
+    public LiteralPacket(int version, int typeId,long value) {
         super(version, typeId);
         this.value = value;
     }
@@ -15,7 +15,7 @@ public class LiteralPacket extends Packet {
     }
 
     @Override
-    public int getValue() {
+    public long getValue() {
         return value;
     }
